wip: run integration tests on nightly only
