+++ .travis.yml
@@ -3,11 +3,13 @@ language: rust
 sudo: required
 
 rust:
-  # - stable
-  # - beta
+  - stable
+  - beta
   - nightly
 
 matrix:
+  allow_failures:
+    - rust: nightly
   fast_finish: true
 
 cache: cargo
