+++ Cargo.toml
@@ -1,5 +1,6 @@
 [package]
 authors = ["Nicholas Dujay <nickdujay@gmail.com>"]
+build = "build.rs"
 categories = [
     "parser-implementations",
     "parsing",
@@ -16,11 +17,8 @@ license = "MIT"
 name = "ecmascript"
 readme = "README.md"
 repository = "https://github.com/dat2/ecmascript"
-version = "0.1.0"
+version = "0.2.0"
 
-[[test]]
-harness = false
-name = "parser_test"
 [badges.codecov]
 branch = "master"
 repository = "dat2/ecmascript"
@@ -33,6 +31,13 @@ status = "actively-developed"
 branch = "master"
 repository = "dat2/ecmascript"
 
+[[test]]
+harness = false
+name = "parser_integration_test"
+
+[build-dependencies]
+rustc_version = "0.2.2"
+
 [dependencies]
 combine = "3.3.2"
 failure = "0.1.1"
@@ -42,6 +47,9 @@ serde = "1.0.70"
 serde_derive = "1.0.70"
 serde_json = "1.0.22"
 unicode-xid = "0.1.0"
+
+[features]
+nightly = []
 
 [lib]
 name = "ecmascript"
