+++ build.rs
@@ -1,0 +1,9 @@
+extern crate rustc_version;
+
+use rustc_version::{version_meta, Channel};
+
+fn main() {
+    if version_meta().unwrap().channel == Channel::Nightly {
+        println!("cargo:rustc-cfg=feature=\"nightly\"");
+    }
+}
